refactor(racing) : Racing 클래스 위치 이동
- Racing 클래스 위치를 racingcar -> racingcar/domain 으로 이동
