+++ src/main/java/racingcar/domain/Racing.java
@@ -1,10 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
-import racingcar.domain.Car;
 import racingcar.utils.Parser;
 
 public class Racing {
